Созданы плавающие лейблы. Доработан JS. сжаты некоторые растровые изоборажения
