+++ css/style.css
@@ -763,6 +763,63 @@ textarea {
   max-height: 255px;
 }
 
+/* Floating label */
+
+.floating-label {
+  width: 100%;
+  box-sizing: border-box;
+  position: relative;
+  display: inline-block;
+  vertical-align: top;
+}
+
+.floating-label input:focus::-webkit-input-placeholder,
+.floating-label textarea:focus::-webkit-input-placeholder {
+  opacity: 0;
+}
+
+.floating-label input:focus::-moz-placeholder,
+.floating-label textarea:focus::-moz-placeholder {
+  opacity: 0;
+}
+
+.floating-label input:focus:-ms-input-placeholder,
+.floating-label textarea:focus:-ms-input-placeholder {
+  opacity: 0;
+}
+
+.floating-label input:focus::placeholder,
+.floating-label textarea:focus::placeholder {
+  opacity: 0;
+}
+
+.floating-label .label-text {
+  position: absolute;
+  bottom: 100%;
+  margin-bottom: 2px;
+  left: 15px;
+  opacity: 0;
+  -webkit-transform: translateY(10px);
+  transform: translateY(10px);
+  font-size: 11px;
+  line-height: 13px;
+  color: #999999;
+  -webkit-transition: opacity 0.1s, color 0.15s, -webkit-transform 0.15s;
+  transition: opacity 0.1s, color 0.15s, -webkit-transform 0.15s;
+  transition: transform 0.15s, opacity 0.1s, color 0.15s;
+  transition: transform 0.15s, opacity 0.1s, color 0.15s, -webkit-transform 0.15s;
+}
+
+.floating-label--active .label-text {
+  opacity: 1;
+  -webkit-transform: translateY(0);
+  transform: translateY(0);
+}
+
+.floating-label--focusing .label-text {
+  color: #5b9ddf;
+}
+
 /* Custom control */
 
 .custom-control {
@@ -2173,7 +2230,7 @@ input[id="slide-3"]:checked ~ .site-wrapper label[for="slide-3"] {
   display: flex;
 }
 
-.subscribe-form input[type="email"] {
+.subscribe-form .floating-label {
   display: block;
   -webkit-box-flex: 1;
   -ms-flex: 1;
@@ -2205,6 +2262,7 @@ input[id="slide-3"]:checked ~ .site-wrapper label[for="slide-3"] {
   bottom: 0;
   left: 0;
   background-color: #d7dfd8;
+  background: #d7dfd8 url("../img/bg-contacts-map.png") no-repeat center;
 }
 
 .contacts-info {
