+++ gulpfile.js
@@ -5,6 +5,8 @@ var cssmin = require('gulp-cssmin');
 var rename = require('gulp-rename');
 var plumber = require('gulp-plumber');
 var csscomb = require('gulp-csscomb');
+var uglify = require('gulp-uglify');
+var pump = require('pump');
 
 gulp.task('css', function() {
   return gulp
@@ -22,6 +24,20 @@ gulp.task('css', function() {
     .pipe(gulp.dest('css/'))
 });
 
-gulp.task('watch-css', ['css'], function() {
+gulp.task('compress-js', function(cb) {
+  pump([
+          gulp.src(['js/script.js', 'js/modal.js', 'js/map.js']),
+          uglify({
+            mangle: true
+          }),
+          rename({suffix: '.min'}),
+          gulp.dest('js/')
+      ],
+      cb
+    );
+});
+
+gulp.task('watch-project', ['css', 'compress-js'], function() {
   gulp.watch('css/style-dev.css', ['css']);
+  gulp.watch('js/script.js', ['compress-js']);
 });
